feat(ui): refactor tag filtering into a new TagFilterPopover component
