+++ frontend/components/common/explore/ExploreContent.tsx
@@ -1,15 +1,14 @@
 'use client';
 
-import {Input} from '@/components/ui/input';
 import {Badge} from '@/components/ui/badge';
 import {Button} from '@/components/ui/button';
-import {Popover, PopoverContent, PopoverTrigger} from '@/components/ui/popover';
 import {Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious} from '@/components/ui/carousel';
 import {InputButton, InputButtonAction, InputButtonProvider, InputButtonSubmit, InputButtonInput} from '@/components/animate-ui/buttons/input';
-import {ChevronLeft, ChevronRight, Search, Compass, Filter, Check, X} from 'lucide-react';
+import {ChevronLeft, ChevronRight, Search, Compass, Filter} from 'lucide-react';
 import {ProjectCard} from '@/components/common/project';
 import {ProjectListItem} from '@/lib/services/project/types';
 import {EmptyState} from '@/components/common/layout/EmptyState';
+import {TagFilterPopover} from '@/components/ui/tag-filter-popover';
 import {motion} from 'motion/react';
 
 interface ExploreContentProps {
@@ -65,23 +64,7 @@ export function ExploreContent({data, LoadingSkeleton, pageSize = 20}: ExploreCo
   } = data;
 
   /** 计算衍生状态 */
-  const filteredTags = (tags || []).filter((tag) =>
-    tag.toLowerCase().includes(tagSearchKeyword.toLowerCase()),
-  );
   const totalPages = Math.ceil(total / pageSize);
-
-  /** 快捷操作 */
-  const handleClearAllTags = () => {
-    (selectedTags || []).forEach(onTagToggle);
-  };
-
-  const handleSelectAllTags = () => {
-    filteredTags.forEach((tag) => {
-      if (!(selectedTags || []).includes(tag)) {
-        onTagToggle(tag);
-      }
-    });
-  };
 
   /** 分页组件 */
   const Pagination = () => {
@@ -179,8 +162,8 @@ export function ExploreContent({data, LoadingSkeleton, pageSize = 20}: ExploreCo
         </InputButtonProvider>
 
         {/* 标签筛选器 */}
-        <Popover open={isTagFilterOpen} onOpenChange={onTagFilterOpenChange}>
-          <PopoverTrigger asChild>
+        <TagFilterPopover
+          trigger={
             <Button
               variant="outline"
               size="sm"
@@ -188,118 +171,15 @@ export function ExploreContent({data, LoadingSkeleton, pageSize = 20}: ExploreCo
             >
               <Filter className="h-4 w-4" />
             </Button>
-          </PopoverTrigger>
-          <PopoverContent className="w-80 p-0" align="end">
-            <div className="p-3 space-y-3">
-              <div className="flex flex-col space-y-2">
-                <div className="flex items-center justify-between">
-                  <span className="text-sm font-medium">标签筛选</span>
-                  {(selectedTags || []).length > 0 && (
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="h-6 px-2 text-xs text-muted-foreground hover:text-destructive"
-                      onClick={handleClearAllTags}
-                    >
-                      清除全部
-                    </Button>
-                  )}
-                </div>
-
-                <div className="relative">
-                  <Search className="absolute left-2 top-2.5 h-3 w-3 text-muted-foreground" />
-                  <Input
-                    placeholder="搜索标签..."
-                    value={tagSearchKeyword}
-                    onChange={(e) => onTagSearchKeywordChange(e.target.value)}
-                    className="pl-7 h-8 text-xs"
-                  />
-                </div>
-              </div>
-
-              <div className="max-h-[220px] overflow-y-auto">
-                {filteredTags.length > 0 ? (
-                  <>
-                    <div className="flex items-center justify-between mb-2 px-1">
-                      <span className="text-xs text-muted-foreground">
-                        {`${filteredTags.length} 个标签${tagSearchKeyword ? '匹配' : ''}`}
-                      </span>
-                      {filteredTags.length > 1 && (
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          className="h-6 px-2 text-xs text-primary"
-                          onClick={handleSelectAllTags}
-                        >
-                          全选
-                        </Button>
-                      )}
-                    </div>
-
-                    <div className="space-y-1 mt-1">
-                      {filteredTags.map((tag) => {
-                        const isSelected = (selectedTags || []).includes(tag);
-                        return (
-                          <div
-                            key={`tag-filter-${tag}`}
-                            className={`flex items-center justify-between p-2 rounded-lg cursor-pointer transition-colors hover:bg-muted ${
-                              isSelected ? 'bg-primary/10 border border-primary/20' : ''
-                            }`}
-                            onClick={() => onTagToggle(tag)}
-                          >
-                            <span className={`text-xs font-medium ${
-                              isSelected ? 'text-primary' : 'text-foreground'
-                            }`}>
-                              {tag}
-                            </span>
-                            {isSelected && (
-                              <Check className="h-3 w-3 text-primary" />
-                            )}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </>
-                  ) : (
-                  <div className="text-center py-6">
-                    <p className="text-xs text-muted-foreground">
-                      {tagSearchKeyword ? '未找到匹配的标签' : '暂无标签'}
-                    </p>
-                  </div>
-                  )}
-              </div>
-
-              {(selectedTags || []).length > 0 && (
-                <div className="pt-2 border-t">
-                  <div className="flex items-center justify-between mb-2">
-                    <span className="text-xs font-medium">已选择的标签</span>
-                    <span className="text-xs text-muted-foreground">
-                      {(selectedTags || []).length} 个标签
-                    </span>
-                  </div>
-                  <div className="flex flex-wrap gap-1 mt-1">
-                    {(selectedTags || []).map((tag) => (
-                      <Badge
-                        key={`selected-tag-${tag}`}
-                        variant="outline"
-                        className="px-2 py-0 h-6 bg-primary/5 text-xs text-primary border-primary/20 flex items-center gap-1"
-                      >
-                        {tag}
-                        <X
-                          className="h-3 w-3 cursor-pointer hover:text-destructive"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            onTagToggle(tag);
-                          }}
-                        />
-                      </Badge>
-                    ))}
-                  </div>
-                </div>
-              )}
-            </div>
-          </PopoverContent>
-        </Popover>
+          }
+          tags={tags}
+          selectedTags={selectedTags}
+          tagSearchKeyword={tagSearchKeyword}
+          isOpen={isTagFilterOpen}
+          onTagToggle={onTagToggle}
+          onTagSearchKeywordChange={onTagSearchKeywordChange}
+          onOpenChange={onTagFilterOpenChange}
+        />
       </motion.div>
 
       {/* 即将开始 */}
